[BugFix] Fix the issue that `__ElementFromBinary` cannot create Elements based on `LazyBundle`.
Currently, `__ElementFromBinary` defaults to get information from the default bundle to create Elements. The framework may need to obtain some information from the `LazyBundle` to create Elements, but `__ElementFromBinary` currently lacks this capability. Therefore, a third parameter will be added to indicate whether it is necessary to obtain the corresponding information from the `LazyBundle`, and which `LazyBundle` to retrieve it from.

issue:m-4676645619
AutoSubmit:True